commit 2

+++ database/migrations/2025_08_20_040731_create_discounts_table.php
@@ -4,31 +4,21 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    /**
-     * Run the migrations.
-     */
-    public function up(): void
-    {
+return new class extends Migration {
+    public function up(): void {
         Schema::create('discounts', function (Blueprint $table) {
             $table->id();
             $table->string('name');
             $table->enum('type', ['percentage', 'fixed']);
-            $table->decimal('value', 10, 2);
+            $table->integer('value');
             $table->date('start_date');
             $table->date('end_date');
             $table->boolean('is_active')->default(true);
             $table->timestamps();
         });
-        
     }
-
-    /**
-     * Reverse the migrations.
-     */
-    public function down(): void
-    {
+    public function down(): void {
         Schema::dropIfExists('discounts');
     }
 };
+
